+++ CHANGELOG.md
@@ -4,13 +4,17 @@
 - RegisterProduct screen implementation.
 - Added Workshop Card atom.
 - Added screen for view Workshop 
+- Added function for searching workshops by name.
+
 ### Fixed
 - ArcaApi get all external collabs
 - Search users
+- Fix snackbar for register new workshops
 
 ## [0.1.0] - 12/11/2025
 
 ### Added
+- Consult Product Batches + Register ViewModel and Screen
 - DeleteOneSupply Button Logic
 - Filter and search bar for users
 - US to add a new supply with an image